Adding application logic and frontend
Setting up the project structure, creating the streamlit frontend and mapping based on sample data. This is functional app where we have 10 users who can login and see their maps and sample KPI. SQL DB to be plugged in to be used as source instead of pandas df, and actual KPIs to be integrated.